@@ -2,17 +2,16 @@
 
 /**
  * A WordPress plugin for developers to write custom migration tasks.
- *
  * @wordpress-plugin
  * Plugin Name: Upgrade Task Runner
  * Plugin URI: https://github.com/thefrosty/wp-upgrade-task-runner
  * Description: A WordPress plugin for developers to write custom migration tasks.
- * Version: 2.4.3
+ * Version: 2.5.0
  * Author: Austin Passy
  * Author URI: https://github.com/thefrosty
  * Requires at least: 5.4
- * Tested up to: 5.5.3
- * Requires PHP: 7.3
+ * Tested up to: 5.8.1
+ * Requires PHP: 7.4
  * phpcs:disable SlevomatCodingStandard.Namespaces.FullyQualifiedGlobalConstants.NonFullyQualified
  * @package TheFrosty\WpUpgradeTaskRunner
  */
@@ -20,7 +19,7 @@
 namespace TheFrosty\WpUpgradeTaskRunner;
 
 const SLUG = 'wp-upgrade-task-runner';
-const VERSION = '2.4.3';
+const VERSION = '2.5.0';
 
 use TheFrosty\WpUpgradeTaskRunner\Cli\DispatchTasks;
 use TheFrosty\WpUpgradeTaskRunner\Upgrade\DbUpgrade;
@@ -36,7 +35,15 @@ $plugin
     ->add(new DisablePluginUpdateCheck())
     ->add($container[ServiceProvider::UPGRADE_PROVIDER])
     ->add($container[ServiceProvider::TASK_LOADER])
-    ->addOnCondition(DispatchTasks::class, 'class_exists', [\WP_CLI_Command::class], 'plugins_loaded', 10, null, [$container])
+    ->addOnConditionDeferred(
+        DispatchTasks::class,
+        fn(): bool => \defined('\WP_CLI') && \WP_CLI && \class_exists('\WP_CLI'),
+        null,
+        'plugins_loaded',
+        'init',
+        10,
+        false,
+        [$container])
     ->addOnHook(DbUpgrade::class, 'admin_menu', null, true)
     ->addOnHook(TaskCountCheck::class, 'admin_menu', null, true);
 
